rename image-card-with-center-text. remove ternary statement for classnames. implement classnames library
