+++ src/assets/scss/icon-card-with-text.scss
@@ -1,0 +1,19 @@
+.icon-card-with-text {
+  p {
+    color: var(--black-500);
+  }
+  &__left {
+    text-align: left;
+  }
+
+  &__border {
+    border: solid 8px var(--black-500);
+    padding: 0.5em;
+  }
+}
+
+.icon-card-with-text-wrapper {
+  @media only screen and (max-width: 1024px) {
+    padding: 1em;
+  }
+}
