--- src/assets/scss/image-card-with-center-text.scss
@@ -1,11 +1,0 @@
-.image-card-center-text-wrapper__short {
-  text-align: center;
-  @media only screen and (max-width: 1024px) {
-    padding: 1em;
-  }
-}
-
-.image-card-center-text-subtitle__short {
-  text-align: center;
-  margin: 0.5em;
-}
