+++ src/assets/scss/layout.scss
@@ -7,7 +7,7 @@
 @import "./card-with-image.scss";
 @import "./values";
 @import "./simple-card.scss";
-@import "./image-card-with-center-text.scss";
+@import "./icon-card-with-text.scss";
 @import "./carousel.scss";
 @import "./events.scss";
 @import "./engage.scss";
